+++ lib/renderer_ascii.dart
@@ -1,0 +1,4 @@
+/// The Canvas renderer
+library cellular_automata.renderers.ascii;
+
+export 'src/renderers/ascii_renderer.dart';